requirements.txt
boto3>=1.9
ruamel.yaml>=0.16.5
pathlib==1.0.1
pyyaml>=5.3.1
kafka-python>=2.0.1
